feat(widget): put the numbers in the message and a generic CTA on the button
