+++ packages/widget/src/i18n/es.json
@@ -100,24 +100,25 @@
     "routeIssue": {
       "otherReasons_one": "",
       "otherReasons_other": "",
+      "hideOtherReasons": "",
+      "lookForReasons": "",
+      "applySuggestion": "",
       "amountTooLow": {
         "title": "",
         "description": "",
         "descriptionUsd": "",
-        "descriptionNoAmount": "",
-        "action": ""
+        "descriptionNoAmount": ""
       },
       "amountTooHigh": {
         "title": "",
         "description": "",
-        "descriptionNoAmount": "",
-        "action": ""
+        "descriptionNoAmount": ""
       },
       "slippageTooTight": {
         "title": "",
         "description": "",
-        "descriptionNoAmount": "",
-        "action": ""
+        "descriptionSuggested": "",
+        "descriptionNoAmount": ""
       },
       "destinationAccountNotReady": {
         "title": "",
@@ -148,9 +149,7 @@
       "pairNotSupported": {
         "title": "",
         "description": ""
-      },
-      "hideOtherReasons": "",
-      "lookForReasons": ""
+      }
     },
     "title": {
       "autoRefuel": "Obtener gas {{chainName}}",
